makefile updates, compiling
